[SPARK-50027][SQL] Move Identity Column SQL parsing code to DataTypeAstBuilder
### What changes were proposed in this pull request?

It moves code parsing Identity Column DDL from AstBuilder to DataTypeAstBuilder.

### Why are the changes needed?

`DataTypeAstBuilder` is intended for parsing code of column definitions.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existent tests cover this.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48543 from zhipengmao-db/zhipengmao-db/id-column-refactor.

Authored-by: zhipeng.mao <zhipeng.mao@databricks.com>
