+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/parser/AstBuilder.scala
@@ -45,7 +45,7 @@ import org.apache.spark.sql.catalyst.trees.TreePattern.PARAMETER
 import org.apache.spark.sql.catalyst.types.DataTypeUtils
 import org.apache.spark.sql.catalyst.util.{CharVarcharUtils, DateTimeUtils, IntervalUtils}
 import org.apache.spark.sql.catalyst.util.DateTimeUtils.{convertSpecialDate, convertSpecialTimestamp, convertSpecialTimestampNTZ, getZoneId, stringToDate, stringToTimestamp, stringToTimestampWithoutTimeZone}
-import org.apache.spark.sql.connector.catalog.{CatalogV2Util, IdentityColumnSpec, SupportsNamespaces, TableCatalog, TableWritePrivilege}
+import org.apache.spark.sql.connector.catalog.{CatalogV2Util, SupportsNamespaces, TableCatalog, TableWritePrivilege}
 import org.apache.spark.sql.connector.catalog.TableChange.ColumnPosition
 import org.apache.spark.sql.connector.expressions.{ApplyTransform, BucketTransform, DaysTransform, Expression => V2Expression, FieldReference, HoursTransform, IdentityTransform, LiteralValue, MonthsTransform, Transform, YearsTransform}
 import org.apache.spark.sql.errors.{DataTypeErrorsBase, QueryCompilationErrors, QueryParsingErrors, SqlScriptingErrors}
@@ -3733,58 +3733,6 @@ class AstBuilder extends DataTypeAstBuilder
     withOrigin(ctx) {
       getDefaultExpression(ctx.expression(), "GENERATED").originalSQL
     }
-
-  /**
-   * Parse and verify IDENTITY column definition.
-   *
-   * @param ctx      The parser context.
-   * @param dataType The data type of column defined as IDENTITY column. Used for verification.
-   * @return Tuple containing start, step and allowExplicitInsert.
-   */
-  protected def visitIdentityColumn(
-      ctx: IdentityColumnContext,
-      dataType: DataType): IdentityColumnSpec = {
-    if (dataType != LongType && dataType != IntegerType) {
-      throw QueryParsingErrors.identityColumnUnsupportedDataType(ctx, dataType.toString)
-    }
-    // We support two flavors of syntax:
-    // (1) GENERATED ALWAYS AS IDENTITY (...)
-    // (2) GENERATED BY DEFAULT AS IDENTITY (...)
-    // (1) forbids explicit inserts, while (2) allows.
-    val allowExplicitInsert = ctx.BY() != null && ctx.DEFAULT() != null
-    val (start, step) = visitIdentityColSpec(ctx.identityColSpec())
-
-    new IdentityColumnSpec(start, step, allowExplicitInsert)
-  }
-
-  override def visitIdentityColSpec(ctx: IdentityColSpecContext): (Long, Long) = {
-    val defaultStart = 1
-    val defaultStep = 1
-    if (ctx == null) {
-      return (defaultStart, defaultStep)
-    }
-    var (start, step): (Option[Long], Option[Long]) = (None, None)
-    ctx.sequenceGeneratorOption().asScala.foreach { option =>
-      if (option.start != null) {
-        if (start.isDefined) {
-          throw QueryParsingErrors.identityColumnDuplicatedSequenceGeneratorOption(ctx, "START")
-        }
-        start = Some(option.start.getText.toLong)
-      } else if (option.step != null) {
-        if (step.isDefined) {
-          throw QueryParsingErrors.identityColumnDuplicatedSequenceGeneratorOption(ctx, "STEP")
-        }
-        step = Some(option.step.getText.toLong)
-        if (step.get == 0L) {
-          throw QueryParsingErrors.identityColumnIllegalStep(ctx)
-        }
-      } else {
-        throw SparkException
-            .internalError(s"Invalid identity column sequence generator option: ${option.getText}")
-      }
-    }
-    (start.getOrElse(defaultStart), step.getOrElse(defaultStep))
-  }
 
   /**
    * Create an optional comment string.
